migrate autotest code for oneflow.Tensor.xxx to test_tensor.py
